Revert "Always include vendor_manifest.xml in PRODUCT_PACKAGES"
This reverts commit 239c5308eb03ac8490ce4e66b62ad9fec3999874.

Reason for revert: This breaks targets that use _only_
DEVICE_MANIFEST_SKUS and DEVICE_MANIFEST_<SKU>_FILES,
since vendor_manifest.xml is non existent on those targets.

The original change assumed it would be possible to add an entry
to PRODUCT_PACKAGES even if not existing, but that's not the case.

In the soong-only case this may work as we'd simply skip the
module installation via m.SkipInstall(), but when generating
Android.mk, as vintf_data does, that's not possible.

Test: m on qualcomm device which only contains SKU specific manifests
Change-Id: I3e4e8f9087640608643cc726bcfa0658467745eb
Signed-off-by: NRanjan-17 <nalinishranjan05@gmail.com>